+++ ami-block-device-reader/outputs.tf
@@ -1,0 +1,4 @@
+// Block devices
+output "block_devices" {
+  value = local.block_devices
+}